requirements.txt
PyQt6>=6.2.0
PyQt6-Qt6>=6.2.0
requests>=2.28.0
pillow>=9.0.0
anthropic>=0.2.0
python-dotenv>=0.20.0
google-cloud-vision>=3.0.0
google-cloud-videointelligence>=2.7.0
opencv-python>=4.6.0